+++ src/completions/index.ts
@@ -1,1 +1,2 @@
+export * from './libraries';
 export * from './std_packages';