+++ src/extension.ts
@@ -1,6 +1,10 @@
 import { ExtensionContext } from 'vscode';
-import { VhdlStdPackageCompletionItemProvider } from './completions';
+import {
+    VhdlLibraryCompletionItemProvider,
+    VhdlStdPackageCompletionItemProvider
+} from './completions';
 
 export function activate(context: ExtensionContext) {
+    context.subscriptions.push(VhdlLibraryCompletionItemProvider);
     context.subscriptions.push(VhdlStdPackageCompletionItemProvider);
 }